requirements.txt
torch
torchvision
torchaudio
numpy
pandas
scikit-learn
Pillow
matplotlib
tqdm
 
